refactor(core): drop element-prefixed module names, relocate transform export, tidy element internals
Pure renames/moves/visibility changes — no behaviour change:
- Rename core/element-interpolation.ts -> interpolation.ts and
  core/element-transform.ts -> transform.ts (element prefix dropped).
- Export TransformTarget/applyElementTransform from the core barrel
  (core/index.ts) instead of re-exporting them through element.ts;
  MatrixTransformTarget stays internal. context.ts imports
  applyElementTransform from core/transform directly.
- Consolidate the world-transform API to a single public Element.getWorldTransform()
  method (drop the $ prefix and the redundant free-function wrapper).
- Rename the now-public Element.getComputedValue() (was protected getComputedStateValue).
- Replace scene.ts's #needsRender JS-private field with a TypeScript private
  _needsRender (# fields break reactive-framework proxying).
- Move elements/polyline-renderers.ts -> math/polyline.ts (renderers are math,
  not an element); math -> context imports stay type-only. Public API unchanged
  (still exported via the @ripl/core root barrel).
