+++ packages/core/src/math/index.ts
@@ -1,5 +1,6 @@
 export * from './constants';
 export * from './geometry';
 export * from './matrix';
+export * from './polyline';
 export * from './structs';
 export * from './types';
